requirements.txt
boto3==1.11.14
# python-pptx==0.6.18
python-pptx
requests
mysql-connector-python